Day 23: Switch from array of usize to array of u32, which is significantly faster

+++ 2020/day-23/src/main.rs
@@ -19,7 +19,7 @@ fn main() {
     let mut tail = 0;
     for value in args.value_of("INPUT").unwrap().chars().map(|character| {
         String::from(character)
-            .parse::<usize>()
+            .parse::<u32>()
             .expect("Failed to parse cup as u8")
     }) {
         max = max.max(value);
@@ -27,7 +27,7 @@ fn main() {
             head = value;
         }
         if tail != 0 {
-            next_cup[tail] = value;
+            next_cup[tail as usize] = value;
         }
         tail = value;
     }
@@ -36,12 +36,12 @@ fn main() {
 
     let cup_count = 1_000_000;
     for value in max + 1..=cup_count {
-        next_cup[tail] = value;
+        next_cup[tail as usize] = value;
         tail = value;
     }
 
     // Complete the circular list
-    next_cup[tail] = head;
+    next_cup[tail as usize] = head;
 
     let steps: usize = args.value_of("STEPS").unwrap().parse().unwrap();
 
@@ -50,28 +50,28 @@ fn main() {
         let mut pick_cursor = current;
         let mut picked = [0; 3];
         for pick in &mut picked {
-            pick_cursor = next_cup[pick_cursor];
+            pick_cursor = next_cup[pick_cursor as usize];
             *pick = pick_cursor;
         }
-        next_cup[current] = next_cup[pick_cursor];
+        next_cup[current as usize] = next_cup[pick_cursor as usize];
 
         let mut destination = (current + cup_count - 2) % cup_count + 1;
         while picked.iter().any(|value| *value == destination) {
             destination = (destination + cup_count - 2) % cup_count + 1
         }
 
-        let destination_next = next_cup[destination];
-        next_cup[destination] = picked[0];
-        next_cup[picked[picked.len() - 1]] = destination_next;
+        let destination_next = next_cup[destination as usize];
+        next_cup[destination as usize] = picked[0];
+        next_cup[picked[picked.len() - 1] as usize] = destination_next;
 
-        current = next_cup[current];
+        current = next_cup[current as usize];
     }
 
     while current != 1 {
-        current = next_cup[current];
+        current = next_cup[current as usize];
     }
 
-    current = next_cup[current];
+    current = next_cup[current as usize];
 
     /*
     for _ in 0..cup_count - 1 {
@@ -83,8 +83,8 @@ fn main() {
 
     let mut product = 1;
     for _ in 0..2 {
-        product *= current as u64;
-        current = next_cup[current];
+        product *= u64::from(current);
+        current = next_cup[current as usize];
     }
 
     println!("Product: {}", product);
